Work on dedicated server. This svn revision is totally broken, it is part of a complete redesign of the server. There is a lobby thread for the server, and each game has its own game processing thread and a sender thread. Connections can be accepted from different sources, i.e. TCP/SCTP over IPv4, IPv6, all those work well in parallel.

+++ src/net/serverrecvstate.h
@@ -124,7 +124,7 @@ protected:
 
 private:
 
-	u_int16_t m_curUniquePlayerId;
+	u_int32_t m_curUniquePlayerId;
 	static boost::thread_specific_ptr<ServerRecvStateInit>	Ptr;
 };
